+++ demo.env.js
@@ -15,6 +15,15 @@
 // -- Main section -
 
 const env = {
+  // The credentials of the pod. There used by the other servers to identify the
+  // pod entering in contact with them.
+  pod: {
+    authserver: 'the name of the server who can confirm the credentials of the pod',
+    auth: {
+      user: 'the authentication username',
+      password: 'the authentication password',
+    },
+  },
   // Define here the credentials to access your database. This hidden file
   // is declared not to be exported in '.gitignore'.
   db: {
@@ -30,6 +39,7 @@ const env = {
       testdb: 'the test database name',
       user: 'the username with the privileges to access to the database',
       password: 'the username password',
+      timezone: 'set the timezone',
     },
   },
 
@@ -46,7 +56,7 @@ const env = {
         useUnifiedTopology: true,
         useNewUrlParser: true,
         connectTimeoutMS: 30000,
-        keepAlive: true,
+        // keepAlive: true,
       },
     },
     testdb: {
@@ -60,11 +70,12 @@ const env = {
         useUnifiedTopology: true,
         useNewUrlParser: true,
         connectTimeoutMS: 30000,
-        keepAlive: true,
+        // keepAlive: true,
       },
     },
   },
 };
+
 
 // -- Export
 module.exports = env;