+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Updated the project dependencies,
   * Replaced nyc by c8,
   * Renamed the package.json scripts,
+  * Reshuffled test/main.js file and added mongodb test,
   * ...,